requirements.txt
Django>=2.1.0,<2.2.6
django-cors-headers
djangorestframework==3.10.3
Pillow==6.2.1
pytz==2019.3
mysqlclient
# django-mysql